remove callab2018 as submodule
somehow callab2018 got recorded as a submodule without ever being
added to the .gitmodules file, and so submodule updates in this
repo would report an error about the missing path.  for this fix
attempt, I created the path and the submodule entry so I could
remove it with git rm.